Use aRequestPrincipal in CSP Should{Load,Process}
CSP Should{Load,Process} always looks for a DOM object to grab the
principal that initiated the request. But if you're using a sandbox
there is no such object and thus CSP will not work. By using the
request principal, we can pass Cu.Sandbox a principal with a CSP
object to enforce CSP on things evaluated in the Sandbox (namely the
request performed with XHR.)